+++ packages/wrapper-react/CHANGELOG.md
@@ -2,10 +2,10 @@
 
 All notable changes to npm module [@typefox/monaco-editor-react](https://www.npmjs.com/package/@typefox/monaco-editor-react) are documented in this file.
 
-## [4.4.0] - 2024-07-1x
+## [4.4.0] - 2024-07-16
 
 - Update to latest monaco-vscode-api [#707](https://github.com/TypeFox/monaco-languageclient/pull/707)
-  - Updated to `monaco-editor-wrapper@5.4.0`, `monaco-languageclient@8.7.0` and version `7.0.5` of `@codingame/monaco-vscode` packages
+  - Updated to `monaco-editor-wrapper@5.4.0`, `monaco-languageclient@8.7.0` and version `7.0.7` of `@codingame/monaco-vscode` packages
 
 ## [4.3.2] - 2024-07-02